commands: move cat code to cat.rs
